Fix tilbakestill filter når ingen innsatsgruppe
