+++ frontend/mulighetsrommet-veileder-flate/src/views/tiltaksgjennomforing-oversikt/ViewTiltakstypeOversikt.tsx
@@ -1,46 +1,19 @@
-import React from 'react';
-import { Alert, Button, Tag } from '@navikt/ds-react';
+import { Button } from '@navikt/ds-react';
 import { useAtom } from 'jotai';
 import { RESET } from 'jotai/utils';
+import { useErrorHandler } from 'react-error-boundary';
 import Filtermeny from '../../components/filtrering/Filtermeny';
+import { BrukersOppfolgingsenhet } from '../../components/oppfolgingsenhet/BrukerOppfolgingsenhet';
 import TiltaksgjennomforingsTabell from '../../components/tabell/TiltaksgjennomforingsTabell';
 import FilterTags from '../../components/tags/Filtertags';
 import SearchFieldTag from '../../components/tags/SearchFieldTag';
+import { useHentBrukerdata } from '../../core/api/queries/useHentBrukerdata';
+import { useInnsatsgrupper } from '../../core/api/queries/useInnsatsgrupper';
 import { tiltaksgjennomforingsfilter, Tiltaksgjennomforingsfiltergruppe } from '../../core/atoms/atoms';
+import { usePrepopulerFilter } from '../../hooks/usePrepopulerFilter';
 import '../../layouts/TiltaksgjennomforingsHeader.less';
 import Show from '../../utils/Show';
 import './ViewTiltakstypeOversikt.less';
-import { usePrepopulerFilter } from '../../hooks/usePrepopulerFilter';
-import { useHentBrukerdata } from '../../core/api/queries/useHentBrukerdata';
-import { kebabCase } from '../../utils/Utils';
-import { useErrorHandler } from 'react-error-boundary';
-import { useInnsatsgrupper } from '../../core/api/queries/useInnsatsgrupper';
-
-function BrukersOppfolgingsenhet() {
-  const brukerdata = useHentBrukerdata();
-  const brukersOppfolgingsenhet = brukerdata?.data?.oppfolgingsenhet?.navn;
-
-  if (brukerdata?.isLoading) {
-    return null;
-  }
-
-  return brukersOppfolgingsenhet ? (
-    <Tag
-      className={'nav-enhet-tag'}
-      key={'navenhet'}
-      variant={brukersOppfolgingsenhet ? 'info' : 'error'}
-      size="small"
-      data-testid={`${kebabCase('filtertag_navenhet')}`}
-      title="Brukers oppfølgingsenhet"
-    >
-      {brukersOppfolgingsenhet}
-    </Tag>
-  ) : (
-    <Alert key="alert-navenhet" data-testid="alert-navenhet" size="small" variant="error">
-      Klarte ikke hente brukers oppfølgingsenhet
-    </Alert>
-  );
-}
 
 const ViewTiltakstypeOversikt = () => {
   const [filter, setFilter] = useAtom(tiltaksgjennomforingsfilter);
@@ -52,6 +25,12 @@ const ViewTiltakstypeOversikt = () => {
 
   const brukersInnsatsgruppeErIkkeValgt = (gruppe: Tiltaksgjennomforingsfiltergruppe) =>
     gruppe.nokkel !== brukerdata?.data?.innsatsgruppe;
+
+  const skalResetteFilter =
+    (filter.innsatsgrupper.length === 0 && !!brukerdata?.data?.innsatsgruppe) ||
+    filter.search !== '' ||
+    filter.tiltakstyper.length > 0 ||
+    filter.innsatsgrupper.some(brukersInnsatsgruppeErIkkeValgt);
 
   return (
     <div className="tiltakstype-oversikt" id="tiltakstype-oversikt" data-testid="tiltakstype-oversikt">
@@ -78,15 +57,7 @@ const ViewTiltakstypeOversikt = () => {
             }
           />
           <SearchFieldTag />
-          <Show
-            if={
-              !innsatsgrupper.isLoading &&
-              (filter.innsatsgrupper.length === 0 ||
-                filter.innsatsgrupper.some(brukersInnsatsgruppeErIkkeValgt) ||
-                filter.search !== '' ||
-                filter.tiltakstyper.length > 0)
-            }
-          >
+          <Show if={!innsatsgrupper.isLoading && skalResetteFilter}>
             <Button
               size="small"
               variant="tertiary"
